Find new stories based on Story#entry_id
Save #entry_id for stories in the database. Use it to check for new items.
Always fall back to url matching when entry_ids is not defined.

This makes sure that feeds with multiple articles using the same url would be imported.
An example of such feeds are feeds generated by http://appshopper.com. All articles link back to a detail page. Multiple stories about that article can appear in the feed.