Add instructions to install foreman
Since pull request #336 https://github.com/swanson/stringer/pull/336 foreman is no longer part of the Gemfile. This needs installing separately as it's not covered by the bundle install (which uses the Gemfile). Finally bundle exec will no longer work on foreman as it's no longer part of the Gemfile so removed the "bundle exec" prefix for starting the application.