Fix Lint/RescueException
This is done by rescuing `StandardError` instead of `Exception`, which should
still cover most if not all errors raised by the Curb and Feedjira libraries. We
no longer try to rescue fatal errors, like `NoMemoryError`, which is what
RuboCop was complaining about.

An even better fix for this could be to rescue an even narrower set of errors,
but I'll leave that until the next time.