Adding clockwork to recurrently fetch feeds
It will replace Scheduler (for deployment on Heroku), so that no extra step is
needed due to another spawn command being called on unicorn.rb starting
clockwork.

Still on Heroku, Procfile is updated, describing the worker and clock
components (if the user wants a more robust architecture).
This fixes #411

Also, for development environments, it will allow fetching feeds without
having to manually running rake commands.

The default interval for fetching feeds is 10 min and it can be customized
by env var FETCH_INTERVAL.