Specs: add tests for FetchFeeds
This is super janky, grabbing the instance variable out of the instance,
but there doesn't seem to be a great way to test for the underlying
issue, which is that in some cases jobs don't work when passing the
entire record. I thought also about inspecting the job more closely, but
wasn't able to think of a great solution there, either. I'd like to
refactor this class some down the road to not try to juggle both
`@feeds_ids` and `@feeds` variables.