Deps: upgrade ActiveRecord to 6.1
I needed to update the format of the json being returned in the tests.
I'm not sure if it actually ends up changing the result when the dates
are converted to JSON strings, but it doesn't look like anything client
or server relies particularly on the format, so we should be fine.

Also updates the formatting in `schema.rb` by running `rake db:migrate`.
All of the changes look to be cosmetic.