Deps: remove SQLite (#564)
This blocks upgrading to ActiveRecord 6.x, as the query for
`true`/`false` values is different for some reason depending on the
database, and it fails tests with SQLite. The simplest solution in this
case is probably to remove SQLite. The production setup uses Postgres,
so better to have a consistent database setup, anyway. This also frees
us up to take advantage of Postgres specific features.