update deploy button link
For some reason it doesn't work without the `template` param for me.
Maybe I'm using a privacy extension that strips out the `referer`.
Explicitly passing the `template` will provide a more reliable
experience, even if it's a little less flexible, as forks will no longer
be able to be deployed as easily.