Add download link when story has enclosure (#667)
Many podcast and other media feeds have both a `url` and an
`enclosure_url`. For example, [the Welcome to Nightvale feed][wn]. In
these cases, there's a good chance the user might want to get straight
to the enclosure rather than being directed to another site to listen to
the track. This allows us to directly download the enclosure rather than
having to hunt it down somewhere else.

[wn]: http://feeds.nightvalepresents.com/welcometonightvalepodcast