Set timezone in app.rb
This sets `Time.zone` inside `app.rb`. Currently it isn't set, which
causes `delayed_job_active_record` to [go down a deprecated path][dj].
`ActiveRecord#default_timezone` is deprecated and will be removed in
Rails 7.1, so setting the timezone avoids this issue and punts the
problem of an outdated queueing system down the road.

[dj]: https://github.com/collectiveidea/delayed_job_active_record/blob/d65b0f9900f5b0c78c341c7c0209c2d138d64ec5/lib/delayed/backend/active_record.rb#L175