fix debug page
The mocking hid the fact that the migration methods were no longer
available, so I updated the tests to use the actual classes.

`ActiveRecord::Base.connection.migration_context.open` always returns a
new instance, so for now we continue passing it down in order to avoid
needing to use `allow_any_instance_of`.