CodeQuality: remove setup_complete logic (#821)
We set `setup_complete` to `true` immediately after the redirect from
creating the user. The only effect of this is that it makes it so that a
user can only view the tutorial page once, which seems unnecessary.

The setup complete logic actually introduces a slight vulnerability.
Even though it shouldn't be possible for a user to exist without
`setup_complete: true`, if there were somehow a user with
`setup_complete: false`, it would allow anybody to access the entire
site without logging in, as `needs_authentication?` returns `false` when
the user has not completed setup.

This is still a problem *after* this change, where if there is *no*
user, the app is fully accessible, aside from the root ("/"). It works
because none of the records are actually associated with the user. This
is not a problem when there is a user. In this case it will properly
require a login. I'll follow up with tighter restrictions to prevent
open access.