fix N+1 queries on feeds/index
This adds a new relation for `Feed#unread_stories` and a new scope to
easily pull in the unread stories count, and uses them on `feeds/index`
to avoid N+1 queries checking for the presence and count of unread
stories.

In my testing, this significantly improves performance on the page.
There is still a smaller N+1 when checking the status for red feeds, and
rendering partials for each feed adds up quite a bit, too, so we may
benefit from in-lining the partial.