Use env files instead of environment variables declared in docker-compose. Closes #1088 (#1091)
It's best practice to leave secrets and environment variables outside of a docker compose file, because if there are secrets we can't add the file to a VCS.

This PR adds a new entrypoint to the Docker container to create required env variables for postgres and stringer if they don't already exist. It also changes the docker-compose.yml file to use these env files instead of environment variables passed in one by one. I also renamed the containers from postgres and web to stringer-postgres and stringer which are more identifiable in a setup that may have multiple services or applications running in docker.

Finally, I also updated the docker compose docs and upgraded the docker compose file version to 3.

The new instructions for easy setup via docker compose are:

- clone the repo
- `touch .env`
- `docker compose up -d`
- visit `localhost`