Fever API: handle empty string for since_id in items
Read You[1] always sends the since_id query parameter,
even when it doesn't send a value to it. That is the query string
it sends is api=&items&since_id=

This ends up as empty string instead of nil in Rails, so let's use
a presence check instead of just truthiness.

[1]: https://github.com/Ashinch/ReadYou