Fever API: Read You compatibility (#1173)
* Fever API: return all ungrouped feeds in a default group

Some clients like Read You[1] expect all feeds in a group
and do not handle ungrouped feeds at all. This ensures
compatibility with them

[1]: https://github.com/Ashinch/ReadYou

* Fever API: handle empty string for since_id in items

Read You[1] always sends the since_id query parameter,
even when it doesn't send a value to it. That is the query string
it sends is api=&items&since_id=

This ends up as empty string instead of nil in Rails, so let's use
a presence check instead of just truthiness.

[1]: https://github.com/Ashinch/ReadYou