Allow empty name during import and use FeedDiscovery to try to figure the name out if possible (#1266)
While testing stringer locally, I tried to import my feeds from The Old Reader. The OPML generated by The Old Reader doesn't contain the feed names, so the import failed.
I made a change to allow nils and the import went through, but then the feeds had no name and it was hard to manage the feeds.

So I replicated the logic from `Feed::Create` to try to get a proper title for the Feed.

This will make the import slower in the case there's nameless entries, but the result seems to be better. I'm falling back to the url if the call to the feed failed to avoid retrying on dead feeds.